+++ src/platforms/Reddit/Reddit.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs";
 import * as path from "path";
 
-import Folder, { FileGroup } from "../../models/Folder";
+import Folder, { FileGroup } from "../../models/Source";
 
 import Platform from "../../models/Platform";
 import { PlatformId } from "..";
