+++ src/platforms/Twitter/Twitter.ts
@@ -1,4 +1,4 @@
-import Folder, { FileGroup } from "../../models/Folder";
+import Folder, { FileGroup } from "../../models/Source";
 
 import Platform from "../../models/Platform";
 import { PlatformId } from "..";
