+++ src/plugins/ImageSize.ts
@@ -1,4 +1,4 @@
-import { FileGroup, FileInfo } from "../models/Folder";
+import { FileGroup, FileInfo } from "../models/Source";
 
 import Plugin from "../models/Plugin";
 import Post from "../models/Post";
